requirements.txt
# IoTwizz Dependencies
# Core
rich>=13.0.0
prompt-toolkit>=3.0.0

# Serial/Hardware
pyserial>=3.5

# Network/Protocol
paho-mqtt>=1.6.0
scapy>=2.5.0
paramiko>=3.0.0
requests>=2.28.0

# AI/Wizz Mode
google-generativeai>=0.6.0
openai>=1.0.0
anthropic>=0.18.0

# Firmware Analysis
# Note: binwalk must be installed separately (apt install binwalk / brew install binwalk)
